Move adapters into new package
