+++ app/src/main/scala/com/waz/zclient/connect/ConnectRequestAdapter.scala
@@ -24,7 +24,7 @@ import com.waz.log.BasicLogging.LogTag.DerivedLogTag
 import com.waz.model.UserId
 import com.waz.service.ZMessaging
 import com.waz.utils.events.{EventContext, Signal}
-import com.waz.zclient.conversationlist.ConversationListAdapter.Incoming
+import com.waz.zclient.conversationlist.adapters.ConversationListAdapter.Incoming
 import com.waz.zclient.utils.RichView
 import com.waz.zclient.{Injectable, Injector}
